Rebase from 'upstream'

+++ libs/opengl/shaders/triangles-no-light.v.glsl
@@ -1,16 +1,16 @@
 R"XXX(#version 300 es
 
 // VERTEX SHADER: Default shader for MRPT CRenderizable objects
-// Jose Luis Blanco Claraco (C) 2019-2022
+// Jose Luis Blanco Claraco (C) 2019-2023
 // Part of the MRPT project
 
 
 in vec3 position;
 in vec4 vertexColor;
 
-uniform mediump mat4 pmv_matrix;
+uniform highp mat4 pmv_matrix;
 
-out mediump vec4 frag_materialColor;
+out lowp vec4 frag_materialColor;
 
 void main()
 {
